+++ CellularAutomaton/Interfaces/IWall.cs
@@ -6,9 +6,9 @@
     {
         public Vector2i Coords { get; set; }
 
-        public byte Light { get; set; }
+        public int Light { get; set; }
 
-        public byte LightDiffusion { get; set; }
+        public int LightDiffusion { get; set; }
 
         public IWall Copy();
